+++ src/com/lukzar/model/elements/Line.java
@@ -26,8 +26,8 @@ public class Line extends Part {
 
     @Override
     public Line reverse() {
-        double x = Configuration.Piece.WIDTH - startPos.getX();
-        double y = startPos.getY();
+        int x = Configuration.Piece.WIDTH - startPos.getX();
+        int y = startPos.getY();
         return new Line(Point.of(x, y));
     }
